added: [ 1781132 ] Various patches to make it compile on 64 bit archs - fileitem.patch. thanks to seppel
git-svn-id: https://xbmc.svn.sourceforge.net/svnroot/xbmc/trunk@10085 568bbfeb-2a22-0410-94d2-cc84cf5bfa90

+++ XBMC/xbmc/FileItem.cpp
@@ -1443,7 +1443,7 @@ void CFileItemList::Serialize(CArchive& ar)
     ar << (int)m_sortOrder;
     ar << m_bCacheToDisc;
 
-    ar << m_sortDetails.size();
+    ar << (int)m_sortDetails.size();
     for (unsigned int j = 0; j < m_sortDetails.size(); ++j)
     {
       const SORT_METHOD_DETAILS &details = m_sortDetails[j];
